Refactor to have a plugins folder
This refactors and demonstrates how a lot of plugins can be added to
this repository.
