+++ src/plugin/redis/resp_parser.rs
@@ -1,14 +1,11 @@
-use anyhow::Result;
 use nom::{
     branch::alt,
     bytes::complete::{tag, take, take_while},
     character::complete::char,
     IResult,
 };
-use std::{collections::HashMap, str, sync::Arc};
-use tokio::sync::Mutex;
 
-use crate::tun::{Handler, Metrics};
+use std::str;
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct RespValue {
@@ -231,66 +228,4 @@ mod tests {
     //    };
     //    assert_eq!(parse_array(input).unwrap().1, expected);
     //}
-}
-
-pub struct RespHandler {
-    port: u16,
-    key_map: Arc<Mutex<HashMap<u32, RespValue>>>,
-}
-
-impl RespHandler {
-    pub fn new(port: u16) -> Self {
-        RespHandler {
-            port,
-            key_map: Arc::new(Mutex::new(HashMap::new())),
-        }
-    }
-}
-
-impl Handler<RespValue> for RespHandler {
-    async fn port(&self) -> u16 {
-        self.port
-    }
-
-    async fn parse_packet(&self, buf: Vec<u8>) -> Result<RespValue> {
-        let resp = parse_resp(&buf).map_err(|_| anyhow::anyhow!("Failed to parse packet"))?;
-        Ok(resp.1)
-    }
-
-    async fn process(&self, input: RespValue, metrics: Option<Metrics>) -> Result<()> {
-        // Return if none and unpack the metrics
-        if metrics.is_none() {
-            return Ok(());
-        }
-        // We already know that metrics is not None
-        let metrics = metrics.unwrap();
-
-        let mut store = self.key_map.lock().await;
-        if !store.contains_key(&metrics.identifier) {
-            // Check if the identifier exists and save it in the store
-            store.insert(metrics.identifier, input.clone());
-        }
-
-        if let Some(latency) = metrics.latency {
-            let status = if input.to_string().contains("ERR") {
-                "ERR"
-            } else {
-                "OK"
-            };
-            // Print the latency and the key
-            let stored_value = store
-                .get(&metrics.identifier)
-                .ok_or_else(|| anyhow::anyhow!("Failed to get value from store"))?;
-            println!(
-                "Key: {}, Latency: {}ms, Status: {}",
-                stored_value.key.as_ref().unwrap(),
-                latency.as_millis(),
-                status,
-            );
-            // clean up the store
-            store.remove(&metrics.identifier);
-        }
-
-        Ok(())
-    }
 }
